+++ packages/components/modules/__shared__/ActionsOverlay/index.tsx
@@ -5,13 +5,17 @@ import {
   SwipeableDrawer as DefaultSwipeableDrawer,
   IconButton,
   TrashCanIcon,
+  usePopover,
 } from '@baseapp-frontend/design-system'
 
 import { LoadingButton } from '@mui/lab'
 import { Box, Divider, Typography } from '@mui/material'
 import { LongPressCallbackReason, useLongPress } from 'use-long-press'
 
-import { ActionOverlayTooltipContainer, IconButtonContentContainer } from './styled'
+import DefaultHoverOverlay from './DefaultHoverOverlay'
+import ThreeDotsMenuHoverOverlay from './ThreeDotsMenuHoverOverlay'
+import { HOVER_OVERLAY_MODES } from './constants'
+import { IconButtonContentContainer } from './styled'
 import { ActionOverlayProps, LongPressHandler } from './types'
 
 const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
@@ -28,6 +32,7 @@ const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
       ContainerProps = {},
       SwipeableDrawerProps = {},
       SwipeableDrawer = DefaultSwipeableDrawer,
+      hoverOverlayMode = HOVER_OVERLAY_MODES.DEFAULT,
     },
     ref,
   ) => {
@@ -37,6 +42,13 @@ const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
       isLongPressingItem: false,
       shouldOpenItemOptions: false,
     })
+
+    const popover = usePopover()
+
+    const handleClosePopover = () => {
+      setIsHoveringItem(false)
+      popover.onClose()
+    }
 
     const longPressHandlers = useLongPress<HTMLDivElement>(
       (e: React.MouseEvent<HTMLDivElement> | React.TouchEvent<HTMLDivElement>) => {
@@ -70,6 +82,7 @@ const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
     const handleDeleteDialogClose = () => {
       setIsDeleteDialogOpen(false)
       setIsHoveringItem(false)
+      popover.onClose()
     }
 
     const deviceHasHover =
@@ -80,6 +93,7 @@ const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
       handleDeleteItem?.()
       handleLongPressItemOptionsClose()
       setIsHoveringItem(false)
+      popover.onClose()
     }
 
     const renderDeleteDialog = () => (
@@ -139,7 +153,9 @@ const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
                       <Box display="grid" justifySelf="center" height="min-content">
                         {icon}
                       </Box>
-                      <Typography variant="body2">{label}</Typography>
+                      <Typography variant="body2" color="text.primary">
+                        {label}
+                      </Typography>
                     </IconButtonContentContainer>
                   </IconButton>
                 )
@@ -170,44 +186,38 @@ const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
       }
 
       if (deviceHasHover && isHoveringItem) {
-        return (
-          <ActionOverlayTooltipContainer
-            offsetRight={offsetRight}
-            offsetTop={offsetTop}
-            aria-label="actions overlay"
-          >
-            {enableDelete && (
-              <IconButton
-                onClick={handleDeleteDialogOpen}
-                disabled={isDeletingItem}
-                aria-label="delete item"
-              >
-                <TrashCanIcon />
-              </IconButton>
-            )}
-            {actions?.map(({ label, icon, onClick, disabled, hasPermission, closeOnClick }) => {
-              if (!hasPermission) return null
+        if (hoverOverlayMode === HOVER_OVERLAY_MODES.DEFAULT) {
+          return (
+            <DefaultHoverOverlay
+              {...{
+                offsetRight,
+                offsetTop,
+                enableDelete,
+                isDeletingItem,
+                handleDeleteDialogOpen,
+                actions,
+                handleLongPressItemOptionsClose,
+              }}
+            />
+          )
+        }
 
-              const handleClick = () => {
-                onClick?.()
-                if (closeOnClick) {
-                  handleLongPressItemOptionsClose()
-                }
-              }
-
-              return (
-                <IconButton
-                  key={label}
-                  onClick={handleClick}
-                  disabled={disabled}
-                  aria-label={label}
-                >
-                  {icon}
-                </IconButton>
-              )
-            })}
-          </ActionOverlayTooltipContainer>
-        )
+        if (hoverOverlayMode === HOVER_OVERLAY_MODES.THREE_DOTS_MENU) {
+          return (
+            <ThreeDotsMenuHoverOverlay
+              {...{
+                offsetRight,
+                offsetTop,
+                enableDelete,
+                isDeletingItem,
+                handleDeleteDialogOpen,
+                actions,
+                handleClosePopover,
+                popover,
+              }}
+            />
+          )
+        }
       }
       return <div />
     }
@@ -216,9 +226,16 @@ const ActionsOverlay = forwardRef<HTMLDivElement, ActionOverlayProps>(
       <Box
         ref={ref}
         onMouseEnter={() => setIsHoveringItem(true)}
-        onMouseLeave={() => setIsHoveringItem(false)}
+        onMouseLeave={handleClosePopover}
         position="relative"
         {...longPressHandlers()}
+        {...(hoverOverlayMode === HOVER_OVERLAY_MODES.THREE_DOTS_MENU && {
+          sx: {
+            width: '100%',
+            display: 'flex',
+            justifyContent: 'flex-end',
+          },
+        })}
         {...ContainerProps}
       >
         {renderDeleteDialog()}