+++ packages/components/modules/messages/MessagesList/MessagesGroup/UserMessage/MessageItem/types.ts
@@ -1,6 +1,6 @@
 import { BoxProps } from '@mui/material'
 
-import { MessageItemFragment$key } from '../../../../../__generated__/MessageItemFragment.graphql'
+import { MessageItemFragment$key } from '../../../../../../__generated__/MessageItemFragment.graphql'
 
 export interface MessageItemProps {
   messageRef: MessageItemFragment$key