+++ packages/components/modules/profiles/Members/MemberListItem/types.ts
@@ -11,4 +11,5 @@ export interface MemberListItemProps {
   nextMember: MemberItemFragment$key | null | undefined
   MemberItemComponent: FC<MemberItemProps>
   memberItemComponentProps?: Partial<MemberItemProps>
+  searchQuery?: string
 }